Create ClarinWorkspaceItemRestRepositoryIT for our WorkspaceItem tests. Created integration tests for maintaining clarin license in the item's metadata and workspace item bitstream.
