+++ dspace-server-webapp/src/test/java/org/dspace/app/rest/WorkspaceItemRestRepositoryIT.java
@@ -122,10 +122,6 @@ public class WorkspaceItemRestRepositoryIT extends AbstractControllerIntegration
     private ItemService itemService;
     @Autowired
     private ConfigurationService configurationService;
-    @Autowired
-    private ClarinLicenseService clarinLicenseService;
-    @Autowired
-    private ClarinLicenseLabelService clarinLicenseLabelService;
 
     private GroupService groupService;
 
@@ -134,11 +130,6 @@ public class WorkspaceItemRestRepositoryIT extends AbstractControllerIntegration
     private Group embargoedGroup2;
     private Group anonymousGroup;
 
-    public static final String REST_SERVER_URL = "http://localhost/api/";
-    public static final String EU_SPONSOR = "EU;Test funding code;Test funding organization;Test funding name;" +
-            "info:eu-repo/grantAgreement/Test funding code";
-    private static final String NON_EU_SPONSOR = "No EU;Test funding code;Test funding organization;Test funding name";
-    private static final String EU_SPONSOR_RELATION = "info:eu-repo/grantAgreement/Test funding code";
 
     @Before
     @Override
@@ -7360,613 +7351,5 @@ public class WorkspaceItemRestRepositoryIT extends AbstractControllerIntegration
                                              bitstream2, ResourcePolicy.TYPE_CUSTOM, Constants.READ, "embargo")
                               )))
                              .andExpect(jsonPath("$.page.totalElements", is(1)));
-    }
-
-    @Test
-    public void uploadFileBiggerThanUploadFileSizeLimit() throws Exception {
-        // Create a big file in the /temp folder
-        String TMP_DIR = System.getProperty("java.io.tmpdir");
-        String FILE_NAME = "test.txt";
-        Long SIZE_IN_BYTES = 2000000000L;
-        File file = null;
-
-        try {
-            file = new File(TMP_DIR, FILE_NAME);
-
-            RandomAccessFile raf = new RandomAccessFile(file, "rw");
-            raf.setLength(SIZE_IN_BYTES);
-            raf.close();
-
-            context.turnOffAuthorisationSystem();
-
-            // A community-collection structure with one parent community and one collection
-            parentCommunity = CommunityBuilder.createCommunity(context)
-                    .withName("Parent Community")
-                    .build();
-            Collection col = CollectionBuilder.createCollection(context, parentCommunity)
-                    .withName("Collection").build();
-
-            WorkspaceItem wItem = WorkspaceItemBuilder.createWorkspaceItem(context, col)
-                    .withTitle("Test WorkspaceItem")
-                    .withIssueDate("2017-10-17")
-                    .withMetadata("local", "bitstream", "redirectToURL", file.getAbsolutePath())
-                    .build();
-
-            context.restoreAuthSystemState();
-            String tokenAdmin = getAuthToken(admin.getEmail(), password);
-
-            List<Operation> addAccessCondition = new ArrayList<Operation>();
-            List<Map<String, String>> accessConditions = new ArrayList<Map<String, String>>();
-
-            Map<String, String> accessCondition1 = new HashMap<String, String>();
-            accessCondition1.put("name", "administrator");
-            accessConditions.add(accessCondition1);
-
-            addAccessCondition.add(new AddOperation("/sections/defaultAC/accessConditions",
-                    accessConditions));
-
-            String patchBody = getPatchContent(addAccessCondition);
-            // add access conditions
-            getClient(tokenAdmin).perform(patch("/api/submission/workspaceitems/" + wItem.getID())
-                            .content(patchBody)
-                            .contentType(MediaType.APPLICATION_JSON_PATCH_JSON))
-                    .andExpect(status().isOk())
-                    .andExpect(jsonPath("$.sections.upload.files[0]"
-                                    + ".metadata['dc.title'][0].value",
-                            is(FILE_NAME)))
-                    .andExpect(jsonPath("$.sections.upload.files[0].sizeBytes",
-                            is(SIZE_IN_BYTES.intValue())));
-        } catch (Exception e) {
-            throw new Exception("Cannot upload file bigger than upload file size limit, " + e.getMessage());
-        } finally {
-            // the big file should be deleted
-            File checkFile = null;
-
-            String shouldDeleteFile = configurationService.getProperty("delete.big.file.after.upload");
-            if (!Objects.isNull(file) && StringUtils.equals("true", shouldDeleteFile)) {
-                checkFile = new File(file.getAbsolutePath());
-                Assert.assertFalse(checkFile.exists());
-            }
-
-            try {
-                // if is not deleted, delete that test big file
-                if (!Objects.isNull(checkFile) && checkFile.exists()) {
-                    FileUtils.forceDelete(file);
-                }
-            } catch (Exception e) {
-                throw new Exception("Cannot delete the file in the end of the test: " + e.getMessage());
-            }
-        }
-    }
-
-    @Test
-    public void patchAddEUFundingMetadataOnItemStillInSubmissionTest() throws Exception {
-        WorkspaceItem witem = this.createSimpleWorkspaceItem();
-
-        List<Operation> addTitle = new ArrayList<Operation>();
-        // create a list of values to use in add operation
-        List<Map<String, String>> titleValues = new ArrayList<Map<String, String>>();
-        Map<String, String> value = new HashMap<String, String>();
-        value.put("value", EU_SPONSOR);
-        titleValues.add(value);
-        addTitle.add(new AddOperation("/sections/traditionalpageone/local.sponsor", titleValues));
-
-        String patchBody = getPatchContent(addTitle);
-
-        // Verify submitter cannot modify metadata via item PATCH. They must use submission forms.
-        String tokenEperson = getAuthToken(admin.getEmail(), password);
-        getClient(tokenEperson).perform(patch("/api/submission/workspaceitems/" + witem.getID())
-                        .content(patchBody)
-                        .contentType(MediaType.APPLICATION_JSON_PATCH_JSON))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$",  WorkspaceItemMatcher.matchItemWithSponsorRelation(witem,
-                        EU_SPONSOR, EU_SPONSOR_RELATION)));
-    }
-
-    @Test
-    public void patchAddNonEUFundingMetadataOnItemStillInSubmissionTest() throws Exception {
-        WorkspaceItem witem = this.createSimpleWorkspaceItem();
-
-        List<Operation> addTitle = new ArrayList<Operation>();
-        // create a list of values to use in add operation
-        List<Map<String, String>> titleValues = new ArrayList<Map<String, String>>();
-        Map<String, String> value = new HashMap<String, String>();
-        value.put("value", NON_EU_SPONSOR);
-        titleValues.add(value);
-        addTitle.add(new AddOperation("/sections/traditionalpageone/local.sponsor", titleValues));
-
-        String patchBody = getPatchContent(addTitle);
-
-        // Verify submitter cannot modify metadata via item PATCH. They must use submission forms.
-        String tokenEperson = getAuthToken(admin.getEmail(), password);
-        getClient(tokenEperson).perform(patch("/api/submission/workspaceitems/" + witem.getID())
-                        .content(patchBody)
-                        .contentType(MediaType.APPLICATION_JSON_PATCH_JSON))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$",  WorkspaceItemMatcher.matchItemWithSponsorRelation(witem,
-                        NON_EU_SPONSOR, null)));
-    }
-
-    @Test
-    public void patchAddMoreNonEUWithEUFundingMetadataOnItemStillInSubmissionTest() throws Exception {
-        WorkspaceItem witem = this.createSimpleWorkspaceItem();
-        String secondNonEUSponsor = "2" + NON_EU_SPONSOR;
-        String secondEUSponsor = "2" + EU_SPONSOR;
-
-        List<Operation> addSponsorOperations = new ArrayList<Operation>();
-
-        // create a list of values to use in add operation
-        List<Map<String, String>> firstSponsorValues = new ArrayList<Map<String, String>>();
-        Map<String, String> firstSponsor = new HashMap<String, String>();
-        firstSponsor.put("value", EU_SPONSOR);
-        firstSponsorValues.add(firstSponsor);
-
-        Map<String, String> secondSponsor = new HashMap<String, String>();
-        secondSponsor.put("value", NON_EU_SPONSOR);
-
-        Map<String, String> thirdSponsor = new HashMap<String, String>();
-        thirdSponsor.put("value", secondNonEUSponsor);
-
-        Map<String, String> fourthSponsor = new HashMap<String, String>();
-        fourthSponsor.put("value", secondEUSponsor);
-
-        addSponsorOperations.add(new AddOperation("/sections/traditionalpageone/local.sponsor", firstSponsorValues));
-        addSponsorOperations.add(new AddOperation("/sections/traditionalpageone/local.sponsor/1", secondSponsor));
-        addSponsorOperations.add(new AddOperation("/sections/traditionalpageone/local.sponsor/2", thirdSponsor));
-        addSponsorOperations.add(new AddOperation("/sections/traditionalpageone/local.sponsor/3", fourthSponsor));
-
-        String patchBody = getPatchContent(addSponsorOperations);
-
-        // Verify submitter cannot modify metadata via item PATCH. They must use submission forms.
-        String tokenEperson = getAuthToken(admin.getEmail(), password);
-        getClient(tokenEperson).perform(patch("/api/submission/workspaceitems/" + witem.getID())
-                        .content(patchBody)
-                        .contentType(MediaType.APPLICATION_JSON_PATCH_JSON))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.sections.traditionalpageone['local.sponsor'][0].value",
-                        is(EU_SPONSOR)))
-                .andExpect(jsonPath("$.sections.traditionalpageone['local.sponsor'][1].value",
-                        is(NON_EU_SPONSOR)))
-                .andExpect(jsonPath("$.sections.traditionalpageone['local.sponsor'][2].value",
-                        is(secondNonEUSponsor)))
-                .andExpect(jsonPath("$.sections.traditionalpageone['local.sponsor'][3].value",
-                        is(secondEUSponsor)))
-                .andExpect(jsonPath("$._embedded.item.metadata['dc.relation'][0].value",
-                        is(EU_SPONSOR_RELATION)))
-                .andExpect(jsonPath("$._embedded.item.metadata['dc.relation'][1].value",
-                        is(EU_SPONSOR_RELATION)));
-    }
-
-    @Test
-    public void patchAddMoreEUWithNonEUFundingMetadataOnItemStillInSubmissionTest() throws Exception {
-        WorkspaceItem witem = this.createSimpleWorkspaceItem();
-        String secondNonEUSponsor = "2" + NON_EU_SPONSOR;
-        String secondEUSponsor = "2" + EU_SPONSOR;
-
-        List<Operation> addSponsorOperations = new ArrayList<Operation>();
-
-        // create a list of values to use in add operation
-        List<Map<String, String>> firstSponsorValues = new ArrayList<Map<String, String>>();
-        Map<String, String> firstSponsor = new HashMap<String, String>();
-        firstSponsor.put("value", NON_EU_SPONSOR);
-        firstSponsorValues.add(firstSponsor);
-
-        Map<String, String> secondSponsor = new HashMap<String, String>();
-        secondSponsor.put("value", EU_SPONSOR);
-
-        Map<String, String> thirdSponsor = new HashMap<String, String>();
-        thirdSponsor.put("value", secondEUSponsor);
-
-        Map<String, String> fourthSponsor = new HashMap<String, String>();
-        fourthSponsor.put("value", secondNonEUSponsor);
-
-        addSponsorOperations.add(new AddOperation("/sections/traditionalpageone/local.sponsor", firstSponsorValues));
-        addSponsorOperations.add(new AddOperation("/sections/traditionalpageone/local.sponsor/1", secondSponsor));
-        addSponsorOperations.add(new AddOperation("/sections/traditionalpageone/local.sponsor/2", thirdSponsor));
-        addSponsorOperations.add(new AddOperation("/sections/traditionalpageone/local.sponsor/3", fourthSponsor));
-
-        String patchBody = getPatchContent(addSponsorOperations);
-
-        // Verify submitter cannot modify metadata via item PATCH. They must use submission forms.
-        String tokenEperson = getAuthToken(admin.getEmail(), password);
-        getClient(tokenEperson).perform(patch("/api/submission/workspaceitems/" + witem.getID())
-                        .content(patchBody)
-                        .contentType(MediaType.APPLICATION_JSON_PATCH_JSON))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.sections.traditionalpageone['local.sponsor'][0].value",
-                        is(NON_EU_SPONSOR)))
-                .andExpect(jsonPath("$.sections.traditionalpageone['local.sponsor'][1].value",
-                        is(EU_SPONSOR)))
-                .andExpect(jsonPath("$.sections.traditionalpageone['local.sponsor'][2].value",
-                        is(secondEUSponsor)))
-                .andExpect(jsonPath("$.sections.traditionalpageone['local.sponsor'][3].value",
-                        is(secondNonEUSponsor)))
-                .andExpect(jsonPath("$._embedded.item.metadata['dc.relation'][0].value",
-                        is(EU_SPONSOR_RELATION)))
-                .andExpect(jsonPath("$._embedded.item.metadata['dc.relation'][1].value",
-                        is(EU_SPONSOR_RELATION)));
-    }
-
-    // replace nonEU - should do not have relation
-    @Test
-    public void patchReplaceNonEUFundingMetadataOnItemStillInSubmissionTest() throws Exception {
-        WorkspaceItem witem = this.createSimpleWorkspaceItem();
-        String updatedSponsor = "updated" + NON_EU_SPONSOR;
-
-        List<Operation> addSponsorOperations = new ArrayList<Operation>();
-        List<Operation> replaceSponsorOperations = new ArrayList<Operation>();
-
-        // create a list of values to use in add operation
-        List<Map<String, String>> firstSponsorValues = new ArrayList<Map<String, String>>();
-        Map<String, String> firstSponsor = new HashMap<String, String>();
-        firstSponsor.put("value", NON_EU_SPONSOR);
-        firstSponsorValues.add(firstSponsor);
-
-        // creating replace operation
-        Map<String, String> updatedSponsorValue = new HashMap<String, String>();
-        updatedSponsorValue.put("value", updatedSponsor);
-
-        addSponsorOperations.add(new AddOperation("/sections/traditionalpageone/local.sponsor",
-                firstSponsorValues));
-        replaceSponsorOperations.add(new ReplaceOperation("/sections/traditionalpageone/local.sponsor/0",
-                updatedSponsorValue));
-
-        String patchBody = getPatchContent(addSponsorOperations);
-        String updateBody = getPatchContent(replaceSponsorOperations);
-
-        String tokenEperson = getAuthToken(admin.getEmail(), password);
-        // Add operation
-        getClient(tokenEperson).perform(patch("/api/submission/workspaceitems/" + witem.getID())
-                        .content(patchBody)
-                        .contentType(MediaType.APPLICATION_JSON_PATCH_JSON))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$",  WorkspaceItemMatcher.matchItemWithSponsorRelation(witem,
-                        NON_EU_SPONSOR, null)))
-                .andExpect(jsonPath("$._embedded.item.metadata['dc.relation'][0].value").doesNotExist());
-
-        // Replace operation
-        getClient(tokenEperson).perform(patch("/api/submission/workspaceitems/" + witem.getID())
-                        .content(updateBody)
-                        .contentType(MediaType.APPLICATION_JSON_PATCH_JSON))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$",  WorkspaceItemMatcher.matchItemWithSponsorRelation(witem,
-                        updatedSponsor, null)))
-                .andExpect(jsonPath("$._embedded.item.metadata['dc.relation'][0].value").doesNotExist());
-    }
-
-    // replace EU - should do not have relation
-    @Test
-    public void patchReplaceEUFundingMetadataOnItemStillInSubmissionTest() throws Exception {
-        WorkspaceItem witem = this.createSimpleWorkspaceItem();
-        String updatedSponsor = "updated" + EU_SPONSOR;
-
-        List<Operation> addSponsorOperations = new ArrayList<Operation>();
-        List<Operation> replaceSponsorOperations = new ArrayList<Operation>();
-
-        // create a list of values to use in add operation
-        List<Map<String, String>> firstSponsorValues = new ArrayList<Map<String, String>>();
-        Map<String, String> firstSponsor = new HashMap<String, String>();
-        firstSponsor.put("value", EU_SPONSOR);
-        firstSponsorValues.add(firstSponsor);
-
-        // creating replace operation
-        Map<String, String> updatedSponsorValue = new HashMap<String, String>();
-        updatedSponsorValue.put("value", updatedSponsor);
-
-        addSponsorOperations.add(new AddOperation("/sections/traditionalpageone/local.sponsor",
-                firstSponsorValues));
-        replaceSponsorOperations.add(new ReplaceOperation("/sections/traditionalpageone/local.sponsor/0",
-                updatedSponsorValue));
-
-        String patchBody = getPatchContent(addSponsorOperations);
-        String updateBody = getPatchContent(replaceSponsorOperations);
-
-        String tokenEperson = getAuthToken(admin.getEmail(), password);
-        // Add operation
-        getClient(tokenEperson).perform(patch("/api/submission/workspaceitems/" + witem.getID())
-                        .content(patchBody)
-                        .contentType(MediaType.APPLICATION_JSON_PATCH_JSON))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$",  WorkspaceItemMatcher.matchItemWithSponsorRelation(witem,
-                        EU_SPONSOR, EU_SPONSOR_RELATION)))
-                .andExpect(jsonPath("$._embedded.item.metadata['dc.relation'][1].value").doesNotExist());
-
-        // Replace operation
-        getClient(tokenEperson).perform(patch("/api/submission/workspaceitems/" + witem.getID())
-                        .content(updateBody)
-                        .contentType(MediaType.APPLICATION_JSON_PATCH_JSON))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$",  WorkspaceItemMatcher.matchItemWithSponsorRelation(witem,
-                        updatedSponsor, EU_SPONSOR_RELATION)))
-                .andExpect(jsonPath("$._embedded.item.metadata['dc.relation'][1].value").doesNotExist());
-    }
-
-    // replace non EU to EU - should have relation
-    @Test
-    public void patchReplaceFromNonEUToEUFundingMetadataOnItemStillInSubmissionTest() throws Exception {
-        WorkspaceItem witem = this.createSimpleWorkspaceItem();
-
-        List<Operation> addSponsorOperations = new ArrayList<Operation>();
-        List<Operation> replaceSponsorOperations = new ArrayList<Operation>();
-
-        // create a list of values to use in add operation
-        List<Map<String, String>> firstSponsorValues = new ArrayList<Map<String, String>>();
-        Map<String, String> firstSponsor = new HashMap<String, String>();
-        firstSponsor.put("value", NON_EU_SPONSOR);
-        firstSponsorValues.add(firstSponsor);
-
-        // creating replace operation
-        Map<String, String> updatedSponsorValue = new HashMap<String, String>();
-        updatedSponsorValue.put("value", EU_SPONSOR);
-
-        addSponsorOperations.add(new AddOperation("/sections/traditionalpageone/local.sponsor",
-                firstSponsorValues));
-        replaceSponsorOperations.add(new ReplaceOperation("/sections/traditionalpageone/local.sponsor/0",
-                updatedSponsorValue));
-
-        String patchBody = getPatchContent(addSponsorOperations);
-        String updateBody = getPatchContent(replaceSponsorOperations);
-
-        String tokenEperson = getAuthToken(admin.getEmail(), password);
-        // Add operation
-        getClient(tokenEperson).perform(patch("/api/submission/workspaceitems/" + witem.getID())
-                        .content(patchBody)
-                        .contentType(MediaType.APPLICATION_JSON_PATCH_JSON))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$",  WorkspaceItemMatcher.matchItemWithSponsorRelation(witem,
-                        NON_EU_SPONSOR, null)))
-                .andExpect(jsonPath("$._embedded.item.metadata['dc.relation'][0].value").doesNotExist());
-
-        // Replace operation
-        getClient(tokenEperson).perform(patch("/api/submission/workspaceitems/" + witem.getID())
-                        .content(updateBody)
-                        .contentType(MediaType.APPLICATION_JSON_PATCH_JSON))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$",  WorkspaceItemMatcher.matchItemWithSponsorRelation(witem,
-                        EU_SPONSOR, EU_SPONSOR_RELATION)))
-                .andExpect(jsonPath("$._embedded.item.metadata['dc.relation'][1].value").doesNotExist());
-    }
-
-    // replace EU to non EU - should do not have relation
-    @Test
-    public void patchReplaceFromRUToNonEUFundingMetadataOnItemStillInSubmissionTest() throws Exception {
-        WorkspaceItem witem = this.createSimpleWorkspaceItem();
-
-        List<Operation> addSponsorOperations = new ArrayList<Operation>();
-        List<Operation> replaceSponsorOperations = new ArrayList<Operation>();
-
-        // create a list of values to use in add operation
-        List<Map<String, String>> firstSponsorValues = new ArrayList<Map<String, String>>();
-        Map<String, String> firstSponsor = new HashMap<String, String>();
-        firstSponsor.put("value", EU_SPONSOR);
-        firstSponsorValues.add(firstSponsor);
-
-        // creating replace operation
-        Map<String, String> updatedSponsorValue = new HashMap<String, String>();
-        updatedSponsorValue.put("value", NON_EU_SPONSOR);
-
-        addSponsorOperations.add(new AddOperation("/sections/traditionalpageone/local.sponsor", firstSponsorValues));
-        replaceSponsorOperations.add(new ReplaceOperation("/sections/traditionalpageone/local.sponsor/0",
-                updatedSponsorValue));
-
-        String patchBody = getPatchContent(addSponsorOperations);
-        String updateBody = getPatchContent(replaceSponsorOperations);
-
-        String tokenEperson = getAuthToken(admin.getEmail(), password);
-        // Add operation
-        getClient(tokenEperson).perform(patch("/api/submission/workspaceitems/" + witem.getID())
-                        .content(patchBody)
-                        .contentType(MediaType.APPLICATION_JSON_PATCH_JSON))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$",  WorkspaceItemMatcher.matchItemWithSponsorRelation(witem,
-                        EU_SPONSOR, EU_SPONSOR_RELATION)))
-                .andExpect(jsonPath("$._embedded.item.metadata['dc.relation'][1].value").doesNotExist());
-
-        // Replace operation
-        getClient(tokenEperson).perform(patch("/api/submission/workspaceitems/" + witem.getID())
-                        .content(updateBody)
-                        .contentType(MediaType.APPLICATION_JSON_PATCH_JSON))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$",  WorkspaceItemMatcher.matchItemWithSponsorRelation(witem,
-                        NON_EU_SPONSOR, null)))
-                .andExpect(jsonPath("$._embedded.item.metadata['dc.relation'][1].value").doesNotExist());
-    }
-
-    // add EU, add nonEU, changee EU to nonEU, change nonEU to EU
-    @Test
-    public void patchAddMoreEUAndNonEUAndReplaceEachOtherOnItemStillInSubmissionTest() throws Exception {
-        WorkspaceItem witem = this.createSimpleWorkspaceItem();
-        String secondEuSponsor = "2" + EU_SPONSOR;
-
-        List<Operation> addSponsorOperations = new ArrayList<Operation>();
-        List<Operation> replaceSponsorOperations = new ArrayList<Operation>();
-
-        // create a list of values to use in add operation
-        List<Map<String, String>> firstSponsorValues = new ArrayList<Map<String, String>>();
-        Map<String, String> firstSponsor = new HashMap<String, String>();
-        firstSponsor.put("value", EU_SPONSOR);
-        firstSponsorValues.add(firstSponsor);
-
-        Map<String, String> secondSponsor = new HashMap<String, String>();
-        secondSponsor.put("value", NON_EU_SPONSOR);
-
-        // creating replace operation
-        Map<String, String> updatedSponsorValue = new HashMap<String, String>();
-        updatedSponsorValue.put("value", secondEuSponsor);
-
-        addSponsorOperations.add(new AddOperation("/sections/traditionalpageone/local.sponsor", firstSponsorValues));
-        addSponsorOperations.add(new AddOperation("/sections/traditionalpageone/local.sponsor/1", secondSponsor));
-        replaceSponsorOperations.add(new ReplaceOperation("/sections/traditionalpageone/local.sponsor/1",
-                updatedSponsorValue));
-
-        String patchBody = getPatchContent(addSponsorOperations);
-        String updateBody = getPatchContent(replaceSponsorOperations);
-
-        String tokenEperson = getAuthToken(admin.getEmail(), password);
-        // Add operation
-        getClient(tokenEperson).perform(patch("/api/submission/workspaceitems/" + witem.getID())
-                        .content(patchBody)
-                        .contentType(MediaType.APPLICATION_JSON_PATCH_JSON))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.sections.traditionalpageone['local.sponsor'][0].value",
-                        is(EU_SPONSOR)))
-                .andExpect(jsonPath("$.sections.traditionalpageone['local.sponsor'][1].value",
-                        is(NON_EU_SPONSOR)))
-                .andExpect(jsonPath("$._embedded.item.metadata['dc.relation'][0].value",
-                        is(EU_SPONSOR_RELATION)))
-                .andExpect(jsonPath("$._embedded.item.metadata['dc.relation'][1].value").doesNotExist());
-
-        // Replace operation
-        getClient(tokenEperson).perform(patch("/api/submission/workspaceitems/" + witem.getID())
-                        .content(updateBody)
-                        .contentType(MediaType.APPLICATION_JSON_PATCH_JSON))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.sections.traditionalpageone['local.sponsor'][0].value",
-                        is(EU_SPONSOR)))
-                .andExpect(jsonPath("$.sections.traditionalpageone['local.sponsor'][1].value",
-                        is(secondEuSponsor)))
-                .andExpect(jsonPath("$._embedded.item.metadata['dc.relation'][0].value",
-                        is(EU_SPONSOR_RELATION)))
-                .andExpect(jsonPath("$._embedded.item.metadata['dc.relation'][1].value",
-                        is(EU_SPONSOR_RELATION)));
-    }
-
-    // add EU, add nonEU, changee EU to nonEU, change nonEU to EU
-    @Test
-    public void patchAddMoreNonEUAndEUAndReplaceEachOtherOnItemStillInSubmissionTest() throws Exception {
-        WorkspaceItem witem = this.createSimpleWorkspaceItem();
-        String secondNonEuSponsor = "2" + NON_EU_SPONSOR;
-
-        List<Operation> addSponsorOperations = new ArrayList<Operation>();
-        List<Operation> replaceSponsorOperations = new ArrayList<Operation>();
-
-        // create a list of values to use in add operation
-        List<Map<String, String>> firstSponsorValues = new ArrayList<Map<String, String>>();
-        Map<String, String> firstSponsor = new HashMap<String, String>();
-        firstSponsor.put("value", NON_EU_SPONSOR);
-        firstSponsorValues.add(firstSponsor);
-
-        Map<String, String> secondSponsor = new HashMap<String, String>();
-        secondSponsor.put("value", EU_SPONSOR);
-
-        // creating replace operation
-        Map<String, String> updatedSponsorValue = new HashMap<String, String>();
-        updatedSponsorValue.put("value", secondNonEuSponsor);
-
-        addSponsorOperations.add(new AddOperation("/sections/traditionalpageone/local.sponsor", firstSponsorValues));
-        addSponsorOperations.add(new AddOperation("/sections/traditionalpageone/local.sponsor/1", secondSponsor));
-        replaceSponsorOperations.add(new ReplaceOperation("/sections/traditionalpageone/local.sponsor/1",
-                updatedSponsorValue));
-
-        String patchBody = getPatchContent(addSponsorOperations);
-        String updateBody = getPatchContent(replaceSponsorOperations);
-
-        String tokenEperson = getAuthToken(admin.getEmail(), password);
-        // Add operation
-        getClient(tokenEperson).perform(patch("/api/submission/workspaceitems/" + witem.getID())
-                        .content(patchBody)
-                        .contentType(MediaType.APPLICATION_JSON_PATCH_JSON))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.sections.traditionalpageone['local.sponsor'][0].value",
-                        is(NON_EU_SPONSOR)))
-                .andExpect(jsonPath("$.sections.traditionalpageone['local.sponsor'][1].value",
-                        is(EU_SPONSOR)))
-                .andExpect(jsonPath("$._embedded.item.metadata['dc.relation'][0].value",
-                        is(EU_SPONSOR_RELATION)))
-                .andExpect(jsonPath("$._embedded.item.metadata['dc.relation'][1].value").doesNotExist());
-
-        // Replace operation
-        getClient(tokenEperson).perform(patch("/api/submission/workspaceitems/" + witem.getID())
-                        .content(updateBody)
-                        .contentType(MediaType.APPLICATION_JSON_PATCH_JSON))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.sections.traditionalpageone['local.sponsor'][0].value",
-                        is(NON_EU_SPONSOR)))
-                .andExpect(jsonPath("$.sections.traditionalpageone['local.sponsor'][1].value",
-                        is(secondNonEuSponsor)))
-                .andExpect(jsonPath("$._embedded.item.metadata['dc.relation'][0].value").doesNotExist())
-                .andExpect(jsonPath("$._embedded.item.metadata['dc.relation'][1].value").doesNotExist());
-    }
-
-    @Test
-    public void addClarinLicenseToWI() throws Exception {
-        WorkspaceItem witem = this.createSimpleWorkspaceItem();
-        List<Operation> replaceOperations = new ArrayList<Operation>();
-        String clarinLicenseName = "Test Clarin License";
-
-        context.turnOffAuthorisationSystem();
-        // Create test clarin license label and clarin license
-        // create LicenseLabels
-        ClarinLicenseLabel clarinLicenseLabel = ClarinLicenseLabelBuilder.createClarinLicenseLabel(context).build();
-        clarinLicenseLabel.setLabel("CC");
-        clarinLicenseLabel.setExtended(false);
-        clarinLicenseLabel.setTitle("CLL Title1");
-        clarinLicenseLabelService.update(context, clarinLicenseLabel);
-
-        ClarinLicense clarinLicense = ClarinLicenseBuilder.createClarinLicense(context).build();
-        clarinLicense.setConfirmation(0);
-        clarinLicense.setDefinition("CL Definition1");
-        clarinLicense.setRequiredInfo("CL Req1");
-        clarinLicense.setName(clarinLicenseName);
-        // add ClarinLicenseLabels to the ClarinLicense
-        HashSet<ClarinLicenseLabel> clarinLicenseLabels = new HashSet<>();
-        clarinLicenseLabels.add(clarinLicenseLabel);
-        clarinLicense.setLicenseLabels(clarinLicenseLabels);
-        clarinLicenseService.update(context, clarinLicense);
-
-        // creating replace operation
-        Map<String, String> licenseReplaceOpValue = new HashMap<String, String>();
-        licenseReplaceOpValue.put("value", clarinLicenseName);
-        replaceOperations.add(new ReplaceOperation("/" + OPERATION_PATH_LICENSE_RESOURCE,
-                licenseReplaceOpValue));
-
-        context.restoreAuthSystemState();
-        String updateBody = getPatchContent(replaceOperations);
-
-        String tokenEperson = getAuthToken(admin.getEmail(), password);
-        // Add operation
-        getClient(tokenEperson).perform(patch("/api/submission/workspaceitems/" + witem.getID())
-                        .content(updateBody)
-                        .contentType(MediaType.APPLICATION_JSON_PATCH_JSON))
-                .andExpect(status().isOk());
-
-        getClient(tokenEperson).perform(get("/api/submission/workspaceitems/" + witem.getID()))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$._embedded.item.metadata['dc.rights'][0].value", is(clarinLicenseName)));
-    }
-
-    @Test
-    public void removeClarinLicenseFromWI() {
-
-    }
-
-    @Test
-    public void updateClarinLicenseInWI() {
-
-    }
-
-    private WorkspaceItem createSimpleWorkspaceItem() {
-        context.turnOffAuthorisationSystem();
-
-        parentCommunity = CommunityBuilder.createCommunity(context)
-                .withName("Parent Community")
-                .build();
-        Community child1 = CommunityBuilder.createSubCommunity(context, parentCommunity)
-                .withName("Sub Community")
-                .build();
-        Collection col1 = CollectionBuilder.createCollection(context, child1)
-                .withName("Collection 1")
-                .build();
-
-        context.setCurrentUser(eperson);
-        WorkspaceItem witem = WorkspaceItemBuilder.createWorkspaceItem(context, col1)
-                .withTitle("Workspace Item 1")
-                .withIssueDate("2017-10-17")
-                .withSubject("ExtraEntry")
-                .build();
-
-        context.restoreAuthSystemState();
-
-        return witem;
     }
 }
